feat(auth): implement hashing with random salt with BCrypt
- fix broken test cases
- delete legacy classes
