+++ src/main/java/woowacourse/auth/domain/Password.java
@@ -1,11 +1,9 @@
 package woowacourse.auth.domain;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import org.springframework.security.crypto.bcrypt.BCrypt;
 
 public class Password {
 
@@ -27,21 +25,8 @@ public class Password {
     }
 
     public EncryptedPassword toEncrypted() {
-        MessageDigest messageDigest = newMessageDigest();
-        messageDigest.update(value.getBytes());
-        return new EncryptedPassword(toEncryptedString(messageDigest.digest()));
-    }
-
-    private MessageDigest newMessageDigest() {
-        try {
-            return MessageDigest.getInstance("SHA-256");
-        } catch (NoSuchAlgorithmException e) {
-            throw new IllegalStateException("암호화 작업을 수행할 수 없습니다.");
-        }
-    }
-
-    private String toEncryptedString(byte[] hashedValue) {
-        return Base64.getEncoder().encodeToString(hashedValue);
+        String randomlyHashedPassword = BCrypt.hashpw(value, BCrypt.gensalt());
+        return new EncryptedPassword(randomlyHashedPassword);
     }
 
     @Override
